Refactore the code partially
