+++ src/model_build.py
@@ -1,9 +1,10 @@
 import pandas as pd
 
-from models.utils import evaluate_cv_search
-from models.model_build import xgboost_train, xgboost_test
+from models.utils import evaluate_cv_search, xgboost_test
+from models.model_build import xgboost_train, get_forecast_by_cluster
 from data.utils import load_data_to_dataframe, get_train_test_data
-from models.utils import get_clusters, get_dataframes_by_cluster, get_forecast_by_cluster
+from models.utils import get_clusters, get_dataframes_by_cluster
+
 
 def main():
     datasets = ["../data/processed/kc_house_data.csv",
@@ -31,10 +32,9 @@ def main():
         dataframe = load_data_to_dataframe(data_path=dataset)
         dataframe = dataframe.drop("date", axis="columns")
         X_train, X_test, y_train, y_test = get_train_test_data(dataframe, ["price"])
-        xgboost_search = xgboost_train(X_train, y_train, n_iter=1)
+        xgboost_search = xgboost_train(X_train, y_train, n_iter=10)
         evaluate_cv_search(xgboost_search)
-        xgboost_test(xgboost_search, X_test, y_test)
-
+        _ = xgboost_test(xgboost_search, X_test, y_test)
 
 
 if __name__ == "__main__":
